Move create/edit exhibit views from user app to core
